BE/app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgi-lifespan==2.1.0
bcrypt==3.2.0
boto3==1.35.46
botocore==1.35.46
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.3
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
joblib==1.4.2
jwt==1.3.1
llvmlite==0.43.0
motor==3.6.0
numba==0.60.0
numpy==2.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.7.0
pymongo==4.9.2
PyMySQL==1.1.1
pynndescent==0.5.13
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
redis==5.0.0
requests==2.32.3
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
umap-learn==0.5.6
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
